+++ CHANGELOG.md
@@ -5,7 +5,10 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.1.11 - Unreleased
+## Version 0.1.12 - Unreleased
+
+
+## Version 0.1.11 - Released 2021-02-24
 
 ### Added
 * Introduced the concept of a bundled dataset
@@ -15,6 +18,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Added `examples` subdirectory for answers to FAQ
 * Added `kwcoco validate` script to check that json and assets exist.
 
+### Changed
+* `CocoDataset.subset` will now only return the videos that are supported by
+  the chosen images.
 
 ## Version 0.1.10 - Released 2021-02-05